% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifCTree.R
\docType{data}
\name{LearnerClassifCTree}
\alias{LearnerClassifCTree}
\title{Classification Conditional Inference Tree Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerClassif}.}
\usage{
LearnerClassifCTree
}
\description{
MISSING.
}
\references{
Breiman, L. (2001).
Random Forests
Machine Learning
\url{https://doi.org/10.1023/A:1010933404324}
}
\keyword{datasets}
