% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrCTree.R
\docType{data}
\name{LearnerRegrCTree}
\alias{LearnerRegrCTree}
\title{Regression Conditional Inference Tree Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerRegr}.}
\usage{
LearnerRegrCTree
}
\description{
MISSING.
}
\references{
Breiman, L. (2001).
Random Forests
Machine Learning
\url{https://doi.org/10.1023/A:1010933404324}
}
\keyword{datasets}
