% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learner.ctree-package}
\alias{mlr3learner.ctree}
\alias{mlr3learner.ctree-package}
\title{mlr3learner.ctree: Wrapper for the ctree learner for use in mlr3}
\description{
mlr3 learner for 'ctree' from the 'partykit' package.
}
\author{
\strong{Maintainer}: Marvin Böcker \email{marvin.boecker@tu-dortmund.de}

}
