% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.partykit-package}
\alias{mlr3learners.partykit}
\alias{mlr3learners.partykit-package}
\title{mlr3learners.partykit: Learners from partykit package for 'mlr3'}
\description{
Adds `ctree()`, and `cforest()` from the partykit
    package to mlr3.
}
\author{
\strong{Maintainer}: Marvin Böcker \email{marvin.boecker@tu-dortmund.de}

Authors:
\itemize{
  \item Marc Becker \email{marcbecker@posteo.de}
  \item Lennart Schneider \email{lennart.sch@web.de} (\href{https://orcid.org/0000-0003-4152-5308}{ORCID})
}

}
