% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifCTree.R
\docType{data}
\name{LearnerClassifCTree}
\alias{LearnerClassifCTree}
\title{Classification Conditional Inference Tree Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerClassif}.}
\usage{
LearnerClassifCTree
}
\description{
A wrapper around ctree (partykit) to use it within mlr3.
}
\references{
Torsten Hothorn, Achim Zeileis (2015).
partykit: A Modular Toolkit for Recursive Partytioning in R.
Journal of Machine Learning Research, 16, 3905-3909.
\url{http://jmlr.org/papers/v16/hothorn15a.html}

Torsten Hothorn, Kurt Hornik and Achim Zeileis (2006).
Unbiased Recursive Partitioning: A Conditional Inference Framework.
Journal of Computational and Graphical Statistics, 15(3), 651--674,
DOI: 10.1198/106186006X133933
}
\keyword{datasets}
