% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.ctree-package}
\alias{mlr3learners.ctree}
\alias{mlr3learners.ctree-package}
\title{mlr3learners.ctree: ctree learner from partykit package for 'mlr3'}
\description{
ctree learner from partykit package for 'mlr3'.
}
\author{
\strong{Maintainer}: Marvin Böcker \email{marvin.boecker@tu-dortmund.de}

Authors:
\itemize{
  \item Marc Becker \email{marcbecker@posteo.de}
}
}
